Refactor on FlightSearchService and move paging & list to Repository layer.
