+++ src/main/java/com/tokigames/api/flights/domain/Flight.java
@@ -19,6 +19,6 @@ public class Flight {
     private FlightType type;
 
     public boolean isLandedInThisCity(String city) {
-        return this.getDeparture().contains(city) || this.getArrival().contains(city);
+        return this.getDeparture().equals(city) || this.getArrival().equals(city);
     }
 }
