+++ src/main/java/com/tokigames/api/flights/port/FlightRepositoryPort.java
@@ -1,13 +1,16 @@
 package com.tokigames.api.flights.port;
 
 import com.tokigames.api.flights.domain.Flight;
+import com.tokigames.api.flights.domain.SortBy;
+import org.springframework.data.domain.Pageable;
 
 import java.util.List;
-import java.util.Set;
 
 public interface FlightRepositoryPort {
 
     void saveFlights(List<Flight> flights);
 
-    Set<Flight> getAllFlights();
+    void deleteAllFlights();
+
+    List<Flight> getFlights(String city, SortBy sortBy, Pageable page);
 }
