+++ src/main/resources/application-test.properties
@@ -1,2 +1,1 @@
 tokigames.flightprovider.api.url=http://localhost:9999
-schedu
