+++ src/test/java/com/tokigames/api/flights/application/FlightSearchServiceQueryTest.java
@@ -1,15 +1,12 @@
 package com.tokigames.api.flights.application;
 
+import com.tokigames.api.flights.adapter.persistence.FlightInMemoryRepository;
 import com.tokigames.api.flights.domain.Flight;
 import com.tokigames.api.flights.domain.FlightType;
 import com.tokigames.api.flights.domain.SortBy;
-import com.tokigames.api.flights.port.FlightRepositoryPort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.internal.util.collections.Sets;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.PageRequest;
 import reactor.core.publisher.Flux;
@@ -19,16 +16,13 @@ import java.util.List;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class FlightSearchServiceQueryTest {
 
-    @Mock
-    private FlightRepositoryPort flightRepositoryPort;
+    private FlightInMemoryRepository flightInMemoryRepository = new FlightInMemoryRepository();
 
-    @InjectMocks
-    private FlightSearchService flightSearchService;
+    private FlightSearchService flightSearchService = new FlightSearchService(flightInMemoryRepository);
 
     @BeforeEach
     public void setup() {
@@ -48,7 +42,8 @@ class FlightSearchServiceQueryTest {
                 LocalDateTime.of(2020, 01, 12, 1, 00),
                 LocalDateTime.of(2020, 01, 12, 3, 00),
                 FlightType.CHEAP);
-        when(flightRepositoryPort.getAllFlights()).thenReturn(Sets.newSet(flight1, flight2, flight3));
+        flightInMemoryRepository.deleteAllFlights();
+        flightInMemoryRepository.saveFlights(List.of(flight1, flight2, flight3));
     }
 
     @Test
